💡refactor: Modifying for Convention
