+++ app/(webhook)/github/_component/webhook-list.tsx
@@ -8,10 +8,10 @@ const WebhookList = ({
 }: WebhookListProps) => {
     return (
         <div className="w-full flex justify-between mt-4 px-4 py-3 bg-GREY-10 rounded-lg">
-            <div className="flex flex-row items-center">
+            <div className="flex-items flex-row">
                 <p className="mr-2">{item.fullName}</p>
-                <div className="w-1 h-[14px] flex-center relative">
-                    <div className="w-[4px] h-[4px] bg-GREY-60 rounded-full absolute top-1.5"></div>
+                <div className="w-1 h-3.5 flex-center relative">
+                    <div className="w-1 h-1 bg-GREY-60 rounded-full absolute top-1.5" />
                 </div>
                 <p className="pl-2">{fullWebhookUrl}</p>
             </div>
