+++ app/repository/_component/register-tab.tsx
@@ -1,7 +1,7 @@
 import { postFetchRepository } from "@/apis/repository";
 import DeliverButton from "@/components/common/button";
 import DeliverInput from "@/components/common/input";
-import { setSessionStorage } from "@/utils/storatge";
+import { setSessionStorage } from "@/utils/storage";
 import { useRouter } from "next/navigation";
 
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
